requirements.txt
glfw==2.6.2
pyopengl==3.1.7
numpy==1.26.1
glm==0.4.4
pillow==10.1.0
pyglm==2.7.1
